scan/requirements.txt
certifi==2020.12.5
numpy==1.20.2
pandas==1.3.5
Pillow==8.2.0
pthflops @ file:///misc/vlgscratch5/PichenyGroup/as15003/pytorch-estimate-flops
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.16.0
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
